+++ packages/orq-rc/RELEASES.md
@@ -249,3 +249,13 @@ Based on:
 - [python v3.1.0-rc.47] packages/orq-rc
 ### Releases
 - [PyPI v3.1.0-rc.47] https://pypi.org/project/orq-ai-sdk/3.1.0-rc.47 - packages/orq-rc
+
+## 2025-02-10 06:49:23
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.487.1 (2.506.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.1.0-rc.48] packages/orq-rc
+### Releases
+- [PyPI v3.1.0-rc.48] https://pypi.org/project/orq-ai-sdk/3.1.0-rc.48 - packages/orq-rc